broke down js into component files, created loading-bar.js to trigger on page load
