+++ components/icon-config.js
@@ -1,0 +1,30 @@
+export const desktopIconArray = [
+    { name: 'home', position: 1, visible: true, empty: false },
+    { name: 'notes', position: 2, visible: true, empty: false },
+    { name: 'Icon3', position: 3, visible: true, empty: true },
+    { name: 'crown', position: 4, visible: true, empty: false },
+    { name: 'snacks', position: 5, visible: true, empty: false },
+    { name: 'photos', position: 6, visible: true, empty: false },
+    { name: 'likes', position: 7, visible: true, empty: false },
+    { name: 'micro-machines', position: 8, visible: true, empty: false },
+    { name: 'Icon9', position: 9, visible: true, empty: true },
+    { name: 'Icon10', position: 10, visible: true, empty: true },
+    { name: 'Icon11', position: 11, visible: true, empty: true },
+    { name: 'Icon12', position: 12, visible: true, empty: true },
+    { name: 'Icon13', position: 13, visible: true, empty: true },
+    { name: 'Icon14', position: 14, visible: true, empty: true },
+    { name: 'Icon15', position: 15, visible: true, empty: true },
+    { name: 'Icon16', position: 16, visible: true, empty: true },
+    { name: 'Icon17', position: 17, visible: true, empty: true },
+    { name: 'Icon18', position: 18, visible: true, empty: true },
+    { name: 'Icon19', position: 19, visible: true, empty: true },
+    { name: 'Icon20', position: 20, visible: true, empty: true },
+    { name: 'Icon21', position: 21, visible: true, empty: true },
+    { name: 'Icon22', position: 22, visible: true, empty: true },
+    { name: 'Icon23', position: 23, visible: true, empty: true },
+    { name: 'Icon24', position: 24, visible: true, empty: true },
+    { name: 'Icon25', position: 25, visible: true, empty: true },
+    { name: 'Icon26', position: 26, visible: true, empty: true },
+    { name: 'Icon27', position: 27, visible: true, empty: true },
+    { name: 'Icon28', position: 28, visible: true, empty: true },
+];
