+++ index.html
@@ -15,6 +15,10 @@
         <!-- Add navigation or other header content as needed -->
     </header>
 
+    <div id="loading-bar">
+        <div id="progress-bar"></div>
+    </div>
+
     <div id="homepage"></div>
 
     <div id="blog-window">
@@ -30,7 +34,7 @@
         <p id="clock"></p>
     </footer>
 
-    <script src="main.js"></script>
+    <script src="main.js" type="module"></script>
 
 </body>
 </html>
